Add MUA_only_flag to power analysis
This adds a new flag "MUA_only_flag" to both dsPlot and dsCalcPower.
This flag tells the code to NOT calculate the power spectra of each
individual population element/cell (aka SUA), but instead to only
calculate the single power spectrum of the averaged activity of a given
variable. This should considerably help with power spectrum analysis
time when simulating large networks, when you mostly ONLY want the
spectrum of the whole thing, NOT the many individual spectra. This has
been tested using most combinations of dsPlot plot enumerations, and
there were no errors found.

Changes to be committed:
	modified:   functions/dsPlot.m
	modified:   functions/internal/dsCalcPower.m